fix(build): extract extractEmbedIds to separate file for testability
Route Handlers cannot have arbitrary named exports (Next.js 15+).
Move extractEmbedIds to extract-embed-ids.ts so it can be exported
for testing while route.ts only exports the GET handler.
